ip-accelerator/requirements.txt
requests==2.31.0
rapidfuzz==3.6.1
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
streamlit==1.30.0
plotly==5.18.0
pyyaml==6.0.1
sqlite-utils==3.36
